common: fix warnings

+++ common/trinary/tryte_ascii.c
@@ -11,7 +11,7 @@
 void ascii_to_trytes(char const *const input, tryte_t *const output) {
   int j = 0, dec = 0, first = 0, second = 0;
 
-  for (int i = 0; input[i]; i++) {
+  for (size_t i = 0; input[i]; i++) {
     dec = input[i];
     first = dec % TRYTE_SPACE;
     second = (dec - first) / TRYTE_SPACE;
@@ -21,7 +21,7 @@ void ascii_to_trytes(char const *const input, tryte_t *const output) {
 }
 
 void trytes_to_ascii(tryte_t const *const input, size_t const input_size, char *const output) {
-  for (int i = 0; i < input_size; i += 2) {
+  for (size_t i = 0; i < input_size; i += 2) {
     output[i / 2] = INDEX_OF_TRYTE(input[i]) + INDEX_OF_TRYTE(input[i + 1]) * TRYTE_SPACE;
   }
 }
